Add files via upload
this project shows developing a text classification model to categorize Amharic texts into two classes: Greeting and Acknowledgment. I have prepared the data by tokenizing and padding the texts to ensure consistent input length for the model. I implemented and trained the model using TensorFlow and Keras, saving it in both .h5 and .keras formats for ease of use. The model was then evaluated with new texts, where I used a label mapping to interpret the predictions accurately.